Create base.abstract_gates
